feat: :sparkles: added the ability to verify commands and close the system corectly

+++ src/motor.c
@@ -1,14 +1,23 @@
 #include "motor_func.h"
 
+// vars globais
 #define PID getpid()
+int fd, command = CLEAR;
 
-int fd, command;
+
+void closeSig(){
+  command = END;
+}
+
 
 int main(int argc, char** argv){
 
+  setbuf(stdout, NULL);
+	setbuf(stdin, NULL);
+
+  // vars de comunicacao
 	prompt cmd;
 	char string[MAX_STRING];
-
   sharedData sharedCmd;
 
 	// select
@@ -16,18 +25,20 @@ int main(int argc, char** argv){
 	fd_set fds;
 	struct timeval timeout;
 
+  // iniciar o motor
 	if(!ini(&fd)){
 	  printf("\n%s\n", getError());
 	  return 1;
 	}
 
+  // handlers de sinais
+	signal(SIGINT, closeSig);
+
 	printf("\n%s⋉  Motor [%d] ⋊%s\n\n>> ", C_MOTOR, PID, C_CLEAR);
 	do{
 
 		// limpar a vars
-    fflush(stdin);
 		FD_ZERO(&fds);
-    command = CLEAR;
     cmd.command[0] = 0;
 		cmd.args[0] = 0;
 
@@ -38,25 +49,64 @@ int main(int argc, char** argv){
 
 		res = select(fd + 1, &fds, NULL, NULL, &timeout);
 
-		if (res == -1) {
+		if (res == -1 && command != END) {
 			printf("Occoreu um erro no select\n");
 			close(fd);
 			unlink(MOTOR_FIFO);
 			return 1;
 
 		}
-		else if (res > 0 && FD_ISSET(0, &fds)) { // ler os comandos do ADMIN
+		else if (res > 0 && FD_ISSET(0, &fds) && command != END) { // ler os comandos do ADMIN
 			scanf("%[^\n]", string);
+      setbuf(stdin, NULL);
 
-      sscanf(string, "%s %s", cmd.command, cmd.args);
+      sscanf(string, "%s %[^\n]", cmd.command, cmd.args);
 
       command = checkCMD(&cmd);
 
-      if(command == CMD_ERROR)
+      switch (command) {
+      case CMD_ERROR:
         printf("\n%s\n",getError());
+      break;
+
+      case USERS:
+        // printUsers();
+      break;
+
+      case KICK:
+        // kick(&cmd);
+      break;
+
+      case BOTS:
+        // printBots();
+      break;
+
+      case BMOV:
+        // rmMv();
+      break;
+
+      case RBM:
+        // rmBm();
+      break;
+
+      case BEGIN:
+        // begin();
+      break;
+
+      case END:
+        // sair
+      break;
+      
+      default:
+        printf("\n%sERRO - nao devia de ter chegado aqui%s\n", C_FATAL_ERROR, C_CLEAR);
+        command = END;
+        break;
+      }
+
+      printf("\n>> ");
 
 		}
-		else if (res > 0 && FD_ISSET(fd, &fds)) { // ler os comandos do FIFO
+		else if (res > 0 && FD_ISSET(fd, &fds)  && command != END) { // ler os comandos do FIFO
 
 			if (read(fd, &sharedCmd, sizeof(sharedData)) > 0) {
 				printf("\n>> ");
@@ -66,6 +116,9 @@ int main(int argc, char** argv){
 
 	} while(command != END);
 
-  printf("\n%s⋉  Bye Bye ⋊%s\n\n>> ", C_MOTOR, C_CLEAR);
+  if(!closeMotor(&fd))
+    printf("%s\nERRO - programa nao foi bem fechado\n%s", C_FATAL_ERROR, C_CLEAR);
+
+  printf("\n%s⋉  Bye Bye ⋊%s\n", C_MOTOR, C_CLEAR);
 	return 0;
 }
